Added QMK firmware with support for Vial. Moved the Micropython code to a separate folder.
