+++ src-macropad/QMK/src/pikku/keymaps/vial/rules.mk
@@ -1,0 +1,3 @@
+VIA_ENABLE= yes
+VIAL_ENABLE= yes
+LTO_ENABLE= yes
