+++ src-macropad/QMK/src/pikku/keymaps/vial/vial.json
@@ -1,0 +1,12 @@
+{
+    "matrix": {
+      "rows": 2,
+      "cols": 3
+    },
+    "layouts": {
+      "keymap": [
+        ["0,0", "0,1", "0,2"],
+        ["1,0", "1,1", "1,2"]
+      ]
+    }
+  }
