+++ src-macropad/QMK/src/pikku/rules.mk
@@ -1,0 +1,5 @@
+OLED_ENABLE= yes
+OLED_DRIVER = ssd1306
+OLED_TRANSPORT = i2c
+I2C1_SDA_PIN = GP2
+I2C1_SCL_PIN = GP3
